requirements.txt
flask>=3.0
pandas>=2.1
python-dateutil>=2.8
psycopg2-binary>=2.9
python-dotenv>=1.0
